Socialapi: move normalizing to its own procedure
